+++ subsys/bluetooth/common/CMakeLists.txt
@@ -7,6 +7,7 @@ zephyr_library_sources(dummy.c)
 zephyr_library_sources(bt_str.c)
 
 zephyr_library_sources_ifdef(CONFIG_BT_RPA   rpa.c)
+zephyr_library_sources_ifdef(CONFIG_BT_LONG_WQ long_wq.c)
 zephyr_library_sources_ifdef(CONFIG_BT_PRIVATE_SHELL bt_shell_private.c)
 
 zephyr_library_link_libraries(subsys__bluetooth)
